+++ Classes/Service/OptimizeImagesFalService.php
@@ -61,7 +61,7 @@ class OptimizeImagesFalService
 
     /**
      * OptimizeImagesFalService constructor.
-     * @param null $config
+     * @param array $config
      * @throws \Exception
      */
     public function __construct($config = null)