+++ CHANGELOG.rst
@@ -6,6 +6,8 @@ master
 ~~~~~~
 
 1) [TASK] Drop travis testing for PHP 5.6 and TYPO3 7.6.
+2) [TASK] TYPO3 9.5 compatibility.
+3) [TASK] ext_localconf.php refactor
 
 2.0.1
 ~~~~~