+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ master
 2) [TASK] TYPO3 9.5 compatibility.
 3) [TASK] ext_localconf.php refactor
 4) [TASK] Drop travis testing for TYPO3 7.6. Add testing for TYPO3 9.5.
+5) [TASK] Increase nimut/testing-framework for TYPO3 9.5 tests
 
 2.0.1
 ~~~~~